requirements/dev.txt
-r base.txt
model-bakery==1.12.0
pytest-django==4.5.2
drf-yasg==1.21.7